Removed private properties.
* Removed private properties.
* Removed update methods.
* Removed __init__ when possible.
* Moved classes to their own files.

Github Issue: #151
